requirements.txt
astroid==1.4.6
colorama==0.3.7
defusedxml==0.5.0
lazy-object-proxy==1.2.2
py==1.4.32
pylint==1.5.6
pytest==3.0.6
PyYAML==3.12
six==1.10.0
wrapt==1.10.8
